Attempted fix for #7 and Syncing improvements
- Changed the connection testing methods to be more specific, adding
support for PMS versions Pre and Post 1.3.0. This will potentially fix
WAN IP resolving issues as it was possible that connections that aren't
working to appear as working.

- Sync improvements. PT will now record how long the client is stuck
buffering when media is played AND when a seek occurs. This value will
then be used to offset seek imes in the future.